Move ImageModal outside of its parent
